Dashboard request numbers are read with PHP's (int) cast, and page offsets are long (#284)
Every hand-written helper in the dashboard that stands in for legacy's `(int)` cast of a request value now reads it with `PhpCast.intval` (D-263). That is PHP 8's cast, held to PHP 8.3.33's own answers by `PhpCastTest`. It reads the leading decimal or exponent number, so `"1e2"` is 100, `"1.9e1"` is 19, `"5abc"` is 5 and `"abc"` is 0.

Before, the helpers did one of two things:
- **Kept only the leading digits**, reading `"1e2"` as 1. These were #249's five: `DashboardListFilters`, `CompanyListFilters`, `Branch`, `Department` and `AdminPayrollController`. `DashboardOrgScope` and `AdminActivitiesController` did the same.
- **Used Java's parse**, refusing `"5abc"` and `"1e2"` and falling back to 0 or a default. These were `AdminRequestsController`, `AdminLeaveBalancesController`, `AdminEmployeeDetailController`, `AdminAttendanceController`, `AdminEmployeesController`, `EmployeeAdminService.contractMonths`, `CompanyForm`, and the guide video, phone country, FAQ and banner forms.

- **Bounds.** A value bound for an `int` field or column is bounded to the `int` range with `Math.clamp`, where PHP's 64-bit value would reach legacy's insert or query. A contract's months are bounded after years become months, so a long duration no longer wraps negative.
- **Offsets.** A page number can now reach `Integer.MAX_VALUE`. `DashboardPage.offsetFor`, its `from` and `to`, and the activities page's offset are now `long`. On `main`, `?page=2147483647` on any list wrapped the offset to a negative `OFFSET`, which the database refuses.
- **`planned_count`.** The column is read as a `long`. A row legacy stored past 2147483647 made MariaDB Connector/J throw `integer overflow`, and the whole workforce planning page answered 500. D-249 gains a correction row; the write side is unchanged.
- **Gate.** `AdminPhpIntCastTest` fails on any `Integer.parseInt`, `Long.parseLong`, `valueOf`, `decode` or `new BigInteger` in `platformadmin`. A digit loop written by hand is not caught. Its one allowed use, `AdminDeviceView.id`, prints an id the database returned.

**Divergences from legacy** (recorded in D-263):
- **Bounds.** A sort order or contract past 2147483647 fails legacy's insert against its `int` column; the port stores the bound. A phone length that large is refused on both sides by the 4 to 15 check before any insert. A page past that bound is `Integer.MAX_VALUE`, which lists nothing, as legacy's page does.
- **A blank or out-of-range year or month.** This behaviour predates this change and is now recorded:
  - Leave balances read a blank, unreadable or non-positive year as this year.
  - The employee detail page reads a blank month or year as this month or year, and a month outside 1 to 12 as this month.
  - Legacy's cast gives 0 or the raw number in each case, which selects nothing.

**Not in this change:** `(float)` casts that stop at the decimal digits in the same way (#280), and numbers bound by Spring as numeric `@RequestParam`s, which answer 400 where legacy's cast reads 0: ids, and payroll's posted month and year (#281).

+++ backend/src/main/java/com/workin/backend/platformadmin/web/DashboardPage.java
@@ -28,9 +28,13 @@ import java.util.List;
  * @param pages   total pages, {@code 0} when there are no rows at all
  * @param from    1-based index of the first row shown, {@code 0} when empty
  * @param to      1-based index of the last row shown
+ *
+ * <p>{@code from}, {@code to} and the offset are {@code long}: a page number reaches
+ * {@code Integer.MAX_VALUE}, and its offset does not fit an {@code int}. Wrapped, it
+ * became a negative {@code OFFSET}, which the database refuses.
  */
 public record DashboardPage<T>(
-		List<T> data, int total, int page, int perPage, int pages, int from, int to) {
+		List<T> data, int total, int page, int perPage, int pages, long from, long to) {
 
 	/** {@code PAGE_SIZE_DEFAULT}. */
 	public static final int SIZE_DEFAULT = 10;
@@ -49,7 +53,7 @@ public record DashboardPage<T>(
 			List<T> rows, int total, int requestedPage, int requestedPerPage) {
 		int perPage = Math.max(1, Math.min(requestedPerPage, SIZE_MAX));
 		int page = Math.max(1, requestedPage);
-		int offset = (page - 1) * perPage;
+		long offset = (page - 1L) * perPage;
 		int pages = total > 0 ? (int) Math.ceil((double) total / perPage) : 0;
 		// The clamp lands here, after the offset above was already taken.
 		int clampedPage = Math.min(page, Math.max(1, pages));
@@ -60,8 +64,8 @@ public record DashboardPage<T>(
 	}
 
 	/** The {@code LIMIT ? OFFSET ?} arguments, from the <em>requested</em> page. */
-	public static int offsetFor(int requestedPage, int perPage) {
-		return (Math.max(1, requestedPage) - 1) * Math.max(1, Math.min(perPage, SIZE_MAX));
+	public static long offsetFor(int requestedPage, int perPage) {
+		return (Math.max(1, requestedPage) - 1L) * Math.max(1, Math.min(perPage, SIZE_MAX));
 	}
 
 	public boolean isEmpty() {
